+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Bug Fixes
 
+* [#2065](https://github.com/bbatsov/rubocop/issues/2065): Handle unary `+` in `ExtraSpacing` cop. ([@jonas054][])
 * [#2275](https://github.com/bbatsov/rubocop/pull/2275): Copy default `Exclude` into `Exclude` lists in `.rubocop_todo.yml`. ([@jonas054][])
 * `Style/IfUnlessModifier` accepts blocks followed by a chained call. ([@lumeet][])
 * [#2261](https://github.com/bbatsov/rubocop/issues/2261): Make relative `Exclude` paths in `$HOME/.rubocop_todo.yml` be relative to current directory. ([@jonas054][])